move enums to seperate files
