+++ src/server/championships.enum.ts
@@ -1,0 +1,5 @@
+export enum Championships {
+  'Midweek League' = '1078',
+  'Mixed Sundays' = '1209',
+  'Summer Challange' = '1366',
+}
