+++ src/server/server-config.dto.ts
@@ -1,16 +1,7 @@
 import { Choice, Param } from '@discord-nestjs/core';
 import { IsAlphanumeric } from 'class-validator';
-
-export enum Championships {
-  // 'Midweek League' = '1078',
-  'Mixed Sundays' = '1209',
-  'Summer Challange' = '1366',
-}
-
-export enum YesOrNo {
-  Yes = '1',
-  No = '0',
-}
+import { Championships } from './championships.enum';
+import { YesOrNo } from './yes-or-no.enum';
 
 export class ServerConfigDto {
   @IsAlphanumeric()
