+++ src/server/yes-or-no.enum.ts
@@ -1,0 +1,4 @@
+export enum YesOrNo {
+  Yes = '1',
+  No = '0',
+}
